+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## To be released
+- Support validating state that contain non-standard JS objects (such as Immutable.js)
+
 ## v0.1.2 (October 31, 2017)
 - Set `lib/index.js` as `main` in package.json